move store classes to new Store namespace
also make them use file-scoped namespace
and get rid of the extra indent
