+++ AcceptanceTest/TestBase.cs
@@ -3,6 +3,7 @@ using QuickFix;
 using System.IO;
 using System.Net;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace AcceptanceTest;
