+++ Examples/Executor/Program.cs
@@ -5,6 +5,7 @@ using System.Runtime.ConstrainedExecution;
 using System.Text;
 using QuickFix;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace Executor
 {
