+++ Examples/SimpleAcceptor/Program.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using QuickFix;
 using QuickFix.Logger;
+using QuickFix.Store;
 using QuickFix.Transport;
 
 namespace SimpleAcceptor
