+++ Examples/TradeClient/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace TradeClient
 {
@@ -32,7 +33,7 @@ namespace TradeClient
             {
                 QuickFix.SessionSettings settings = new QuickFix.SessionSettings(file);
                 TradeClientApp application = new TradeClientApp();
-                QuickFix.IMessageStoreFactory storeFactory = new QuickFix.FileStoreFactory(settings);
+                IMessageStoreFactory storeFactory = new FileStoreFactory(settings);
                 ILogFactory logFactory = new ScreenLogFactory(settings);
                 QuickFix.Transport.SocketInitiator initiator = new QuickFix.Transport.SocketInitiator(application, storeFactory, settings, logFactory);
