+++ QuickFIXn/AbstractInitiator.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Collections.Generic;
 using System;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace QuickFix
 {
