+++ QuickFIXn/Session.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using QuickFix.Fields;
 using QuickFix.Fields.Converters;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace QuickFix
 {
