+++ QuickFIXn/SessionState.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using QuickFix.Logger;
+using QuickFix.Store;
 using MessagesBySeqNum = System.Collections.Generic.Dictionary<ulong, QuickFix.Message>;
 
 namespace QuickFix
