+++ QuickFIXn/Store/IMessageStoreFactory.cs
@@ -1,15 +1,14 @@
-﻿namespace QuickFix
+﻿namespace QuickFix.Store;
+
+/// <summary>
+/// Used by a Session to create a message store implementation
+/// </summary>
+public interface IMessageStoreFactory
 {
     /// <summary>
-    /// Used by a Session to create a message store implementation
+    /// Creates a message store implementation
     /// </summary>
-    public interface IMessageStoreFactory
-    {
-        /// <summary>
-        /// Creates a message store implementation
-        /// </summary>
-        /// <param name="sessionID">the session ID, often used to access session configurations</param>
-        /// <returns>the message store implementation</returns>
-        IMessageStore Create(SessionID sessionID);
-    }
+    /// <param name="sessionId">the session ID, often used to access session configurations</param>
+    /// <returns>the message store implementation</returns>
+    IMessageStore Create(SessionID sessionId);
 }
