+++ QuickFIXn/Transport/SocketInitiator.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Threading;
 using System.IO;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace QuickFix.Transport
 {
