+++ UnitTests/SessionDynamicTest.cs
@@ -9,6 +9,7 @@ using System.Text.RegularExpressions;
 using NUnit.Framework;
 using QuickFix;
 using QuickFix.Logger;
+using QuickFix.Store;
 using QuickFix.Transport;
 
 namespace UnitTests
