+++ UnitTests/SessionStateTest.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using QuickFix.Logger;
+using QuickFix.Store;
 
 namespace UnitTests
 {
